changing types
