+++ openmetadata-ui/src/main/resources/ui/src/axiosAPIs/rolesAPI.ts
@@ -13,7 +13,7 @@
 
 import { AxiosResponse } from 'axios';
 import { Operation } from 'fast-json-patch';
-import { Policy } from 'Models';
+import { Policy } from '../pages/RolesPage/policy.interface';
 import { getURLWithQueryFields } from '../utils/APIUtils';
 import APIClient from './index';
